+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added dynamic Pi startup resource summaries to empty chat screens.
 - Added support for Pi extension `ctx.ui.setFooter(factory)`
 - Added `@` file suggestions to the composer.